fix some type conversion warnings

+++ Source/ALS/Private/AlsCharacter_Actions.cpp
@@ -205,7 +205,7 @@ bool AAlsCharacter::TryStartMantling(const FAlsMantlingTraceSettings& TraceSetti
 
 	static const FName ForwardTraceTag{FString::Printf(TEXT("%hs (Forward Trace)"), __FUNCTION__)};
 
-	auto ForwardTraceStart{CapsuleBottomLocation - ForwardTraceDirection * CapsuleRadius};
+	auto ForwardTraceStart{CapsuleBottomLocation - ForwardTraceDirection * static_cast<double>(CapsuleRadius)};
 	ForwardTraceStart.Z += (TraceSettings.LedgeHeight.X + TraceSettings.LedgeHeight.Y) *
 		0.5f * CapsuleScale - UCharacterMovementComponent::MAX_FLOOR_DIST;
 
@@ -642,7 +642,7 @@ void AAlsCharacter::StartRagdollingImplementation()
 
 		// TODO Find a better solution or wait for a fix in future engine versions.
 
-		static constexpr auto MinSpeedLimit{200.0f};
+		static constexpr auto MinSpeedLimit{200.0};
 
 		RagdollingState.SpeedLimitFrameTimeRemaining = 8;
 		RagdollingState.SpeedLimit = FMath::Max(LocomotionState.Velocity.Size(), MinSpeedLimit);
@@ -768,7 +768,7 @@ void AAlsCharacter::RefreshRagdollingActorTransform(const float DeltaTime)
 			RootBoneHorizontalSpeedSquared > FMath::Square(300.0f) ? UAlsConstants::Spine03BoneName() : UAlsConstants::PelvisBoneName()
 		};
 
-		GetMesh()->AddForce((RagdollTargetLocation - GetMesh()->GetSocketLocation(PullForceSocketName)) * RagdollingState.PullForce,
+		GetMesh()->AddForce((RagdollTargetLocation - GetMesh()->GetSocketLocation(PullForceSocketName)) * static_cast<double>(RagdollingState.PullForce),
 		                    PullForceSocketName, true);
 	}
